# pkg/server/route/docs/swagger.yaml
definitions:
...
  service.HttpRsp_ServiceStep:
    properties:
      args:
        type: object
      cluster_uuid:
        description: pk char(32) cluster.uuid
        type: string
      created:
        type: string
      ended:
        type: string
      method:
        type: string
      name:
        type: string
      partition_date:
        description: pk date
        type: string
      result_filter:
        type: string
      sequence:
        description: pk tinyint
        type: integer
      started:
        type: string
      status:
        type: integer
      summary:
        type: string
      timestamp:
        description: pk datetime(6)
        type: string
      uuid:
        description: pk char(32) service.uuid
        type: string
    type: object
...
  v3.Session:
    properties:
      cluster_uuid:
        type: string
      created:
        type: string
      deleted:
        type: object
      expiration_time:
        type: object
      id:
        description: pk
        type: integer
      issued_at_time:
        type: object
      token:
        type: string
      updated:
        type: object
      uuid:
        description: uuid
        type: string
    type: object
...
info:
  contact:
    email: jaehoon@nexclipper.io
    url: https://nexclipper.io
  description: this is a sudory server.
  title: SUDORY
  version: 0.0.1
paths:
...
  /server/service/{uuid}/step:
    get:
      consumes:
      - application/json
      description: Get Service Steps
      parameters:
      - description: ServiceStep 의 uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/service.HttpRsp_ServiceStep'
            type: array
      security:
      - ServiceAuth: []
      tags:
      - server/service
  /server/service/{uuid}/step/{sequence}:
    get:
      consumes:
      - application/json
      description: Get Service Step
      parameters:
      - description: ServiceStep 의 uuid
        in: path
        name: uuid
        required: true
        type: string
      - description: ServiceStep 의 sequence
        in: path
        name: sequence
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.HttpRsp_ServiceStep'
      security:
      - ServiceAuth: []
      tags:
      - server/service
  /server/service/step:
    get:
      consumes:
      - application/json
      description: Find Service Steps
      parameters:
      - description: query  github.com/NexClipper/sudory/pkg/server/database/vanilla/stmt/README.md
        in: query
        name: q
        type: string
      - description: order  github.com/NexClipper/sudory/pkg/server/database/vanilla/stmt/README.md
        in: query
        name: o
        type: string
      - description: paging github.com/NexClipper/sudory/pkg/server/database/vanilla/stmt/README.md
        in: query
        name: p
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/service.HttpRsp_ServiceStep'
            type: array
      security:
      - ServiceAuth: []
      tags:
      - server/service
  /server/session:
    get:
      consumes:
      - application/json
      description: Find Session
      parameters:
      - description: query  github.com/NexClipper/sudory/pkg/server/database/vanilla/stmt/README.md
        in: query
        name: q
        type: string
      - description: order  github.com/NexClipper/sudory/pkg/server/database/vanilla/stmt/README.md
        in: query
        name: o
        type: string
      - description: paging github.com/NexClipper/sudory/pkg/server/database/vanilla/stmt/README.md
        in: query
        name: p
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v3.Session'
            type: array
      security:
      - ServiceAuth: []
      tags:
      - server/session
  /server/session/{uuid}:
    delete:
      consumes:
      - application/json
      description: Delete a Session
      parameters:
      - description: Session Uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ServiceAuth: []
      tags:
      - server/session
    get:
      consumes:
      - application/json
      description: Get a Session
      parameters:
      - description: Session Uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v3.Session'
      security:
      - ServiceAuth: []
      tags:
      - server/session
...
securityDefinitions:
  ClientAuth:
    description: token for client api
    in: header
    name: x-sudory-client-token
    type: apiKey
  ServiceAuth:
    description: Bearer token for service api
    in: header
    name: Authorization
    type: apiKey
  XAuthToken:
    description: limit for access sudory api
    in: header
    name: x_auth_token
    type: apiKey
swagger: "2.0"